pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smolagents-capsule"
version = "0.1.0"
description = "Capsule sandbox integration for smolagents — run code in isolated WebAssembly sandboxes"
license = { text = "MIT" }
readme = "README.md"
authors = [
    {name = "Capsule Contributors"}
]

classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "capsule-run>=0.6.3",
    "smolagents>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/mavdol/capsule"
Repository = "https://github.com/mavdol/smolagents-capsule"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
]

[tool.hatch.build.targets.wheel]
packages = ["smolagents_capsule"]

[tool.hatch.metadata]
allow-direct-references = true
